% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphicFunctions.R
\name{txt}
\alias{txt}
\title{txt: the upgraded text function}
\usage{
txt(pos = "center", label = NULL)
}
\arguments{
\item{pos}{position of the text, can be center, topleft,
topright, bottomleft, bottomright,}

\item{label}{the actual text to be plotted}
}
\description{
txt makes it easy to add text to the current active
plot.
}

